+++ server/index.js
@@ -2,22 +2,26 @@ import express from 'express';
 import bodyParser from 'body-parser';
 import graphqlHTTP from 'express-graphql';
 import mongoose from 'mongoose';
+import dotenv from 'dotenv';
 import router from './routes/data';
 import schema from './graphql/schema';
-import resolvers from './graphql/resolvers';
+// import resolvers from './graphql/resolvers';
+import authRoot from './graphql/resolvers/auth';
 
-require('dotenv').config();
+console.log('authRoot: ', authRoot);
+
+dotenv.config();
 
 const server = express();
 const PORT = process.env.PORT || 3001;
 const DB = process.env.DB_URL;
 
-console.log('resolvers: ', resolvers);
-
 if (process.env.NODE_ENV !== 'production') {
     console.log('development');
 }
 else console.log('production');
+
+// createUser({ userInput: { "email": "t3@t.t", "password": "test" } }).then(e => console.log(e._doc));
 
 server.use(
     bodyParser.json(),
@@ -28,9 +32,10 @@ server.use(express.static('dist/src'));
 server.use('/routes', router);
 server.use('/graphql', graphqlHTTP({
     schema,
-    rootValue: resolvers,
+    rootValue: authRoot,
     graphiql: true,
 }));
+
 
 mongoose.connect(DB, { useNewUrlParser: true, useUnifiedTopology: true })
     .then(() => console.log('connected to db'))
